+++ apps/mini-rx-angular-demo/src/app/modules/counter/counter/counter.component.ts
@@ -1,4 +1,4 @@
-import { Component } from '@angular/core';
+import { ChangeDetectionStrategy, Component } from '@angular/core';
 import { Observable } from 'rxjs';
 import { CounterStore } from '../state/counter-store.service';
 
@@ -7,6 +7,7 @@ import { CounterStore } from '../state/counter-store.service';
     templateUrl: './counter.component.html',
     styleUrls: ['./counter.component.css'],
     providers: [CounterStore], // The CounterStore is provided for each counter component instance
+    changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class CounterComponent {
     counter$: Observable<number> = this.counterStore.count$;