+++ libs/mini-rx-store-ng/src/lib/component-store.module.ts
@@ -1,0 +1,13 @@
+import { ModuleWithProviders, NgModule } from '@angular/core';
+import { ComponentStoreConfig, configureComponentStores } from 'mini-rx-store';
+
+@NgModule()
+export class ComponentStoreModule {
+    static forRoot(config: ComponentStoreConfig): ModuleWithProviders<ComponentStoreModule> {
+        configureComponentStores(config);
+
+        return {
+            ngModule: ComponentStoreModule,
+        };
+    }
+}