+++ libs/mini-rx-store-ng/src/lib/spec/ng-modules.spec.ts
@@ -3,17 +3,25 @@ import { StoreModule } from '../store.module';
 import {
     Action,
     Actions,
+    createComponentStore,
     createEffect,
     FeatureStore,
+    ImmutableStateExtension,
+    LoggerExtension,
     ofType,
     Reducer,
+    ReduxDevtoolsExtension,
     Store,
     StoreExtension,
+    UndoExtension,
+    _StoreCore,
 } from 'mini-rx-store';
 import { Injectable, NgModule } from '@angular/core';
 import { catchError, map, mergeMap } from 'rxjs/operators';
 import { of } from 'rxjs';
 import { EffectsModule } from '../effects.module';
+import { ComponentStoreModule } from 'mini-rx-store-ng';
+import { NgReduxDevtoolsExtension } from '../ng-redux-devtools.extension';
 
 export const loadAction: Action = {
     type: 'LOAD',
@@ -137,11 +145,19 @@ class CounterFeatureStore extends FeatureStore<CounterState> {
     }
 }
 
+const globalCsExtensions = [new LoggerExtension(), new ImmutableStateExtension()];
+
+const reduxDevToolsExtension = new ReduxDevtoolsExtension({ name: 'Test Redux DevTools' });
+const stateFromReduxDevTools = {
+    someProp: 'someValue',
+};
+
 describe(`Ng Modules`, () => {
     let actions$: Actions;
     let store: Store;
 
     const rootMetaReducerSpy = jest.fn();
+    const configureStoreSpy = jest.spyOn(_StoreCore, 'configureStore');
 
     function rootMetaReducer(reducer: Reducer<any>): Reducer<any> {
         return (state, action) => {
@@ -173,9 +189,12 @@ describe(`Ng Modules`, () => {
                         counter1: { counter: 111 },
                     },
                     metaReducers: [rootMetaReducer],
-                    extensions: [new SomeExtension()],
+                    extensions: [new SomeExtension(), reduxDevToolsExtension],
                 }),
                 Counter5Module,
+                ComponentStoreModule.forRoot({
+                    extensions: globalCsExtensions,
+                }),
             ],
         });
 
@@ -280,6 +299,68 @@ describe(`Ng Modules`, () => {
                 counter4: { counter: 2 },
                 counter5: { counter: 556 },
             });
+        });
+    });
+
+    describe(`ComponentStore`, () => {
+        // Just make sure that the global config is set via the ComponentStoreModule.forRoot static method
+        // For the other aspects of the config we can rely on the ComponentStore tests
+
+        it('should merge global config with local config', () => {
+            const localCsExtensions = [new UndoExtension()];
+
+            const cs = createComponentStore(undefined, { extensions: localCsExtensions });
+
+            expect(cs['extensions'][0]).toBe(localCsExtensions[0]);
+            expect(cs['extensions'][1]).toBe(globalCsExtensions[0]);
+            expect(cs['extensions'][2]).toBe(globalCsExtensions[1]);
+        });
+    });
+
+    describe(`Redux DevTools extension`, () => {
+        it('should initialize NgReduxDevtoolsExtension', () => {
+            const devToolsExtensionFromConfig: NgReduxDevtoolsExtension = configureStoreSpy.mock
+                .calls[0][0]!['extensions']![1] as NgReduxDevtoolsExtension;
+
+            // It would have been nicer to spy on addExtension, but that did not work without refactor of StoreCore (https://medium.com/@DavideRama/mock-spy-exported-functions-within-a-single-module-in-jest-cdf2b61af642)
+            expect(devToolsExtensionFromConfig).toBeInstanceOf(NgReduxDevtoolsExtension);
+            expect(devToolsExtensionFromConfig['options']).toEqual(
+                expect.objectContaining({
+                    name: 'Test Redux DevTools',
+                    traceLimit: 25,
+                })
+            );
+        });
+
+        it('should update the Store state', () => {
+            const devToolsExtensionFromConfig: NgReduxDevtoolsExtension = configureStoreSpy.mock
+                .calls[0][0]!['extensions']![1] as NgReduxDevtoolsExtension;
+
+            const spy = jest.spyOn(_StoreCore.appState, 'set');
+            JSON.parse = jest.fn().mockImplementationOnce((data) => {
+                return data;
+            });
+
+            devToolsExtensionFromConfig['onDevToolsMessage']({
+                type: 'DISPATCH',
+                payload: {
+                    type: 'JUMP_TO_STATE',
+                },
+                state: stateFromReduxDevTools,
+            });
+
+            expect(spy).toHaveBeenCalledTimes(1);
+            expect(spy).toHaveBeenCalledWith(stateFromReduxDevTools);
+
+            spy.mockReset();
+
+            devToolsExtensionFromConfig['onDevToolsMessage']({
+                type: 'NOT_SUPPORTED_TYPE',
+                payload: {},
+                state: {},
+            });
+
+            expect(spy).toHaveBeenCalledTimes(0);
         });
     });
 });