+++ libs/mini-rx-store/src/lib/component-store.ts
@@ -30,7 +30,7 @@ export function _resetConfig() {
     componentStoreConfig = undefined;
 }
 
-export function configureComponentStores(config: { extensions: ComponentStoreExtension[] }) {
+export function configureComponentStores(config: ComponentStoreConfig) {
     if (!componentStoreConfig) {
         componentStoreConfig = config;
         return;
@@ -128,6 +128,15 @@ export class ComponentStore<StateType extends object>
         this.hasUndoExtension
             ? this.dispatch(undo(action))
             : miniRxError(`${this.constructor.name} has no UndoExtension yet.`);
+    }
+
+    override destroy() {
+        if (this.reducer) {
+            // Dispatch an action really just for logging via LoggerExtension
+            // Only dispatch if a reducer exists (if an initial state was provided or setInitialState was called)
+            this.dispatch(createMiniRxAction(MiniRxActionType.DESTROY, csFeatureKey));
+        }
+        super.destroy();
     }
 }