+++ libs/mini-rx-store/CHANGELOG.md
@@ -2,6 +2,29 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [5.0.0-alpha.2](https://github.com/spierala/mini-rx-store/compare/mini-rx-store-5.0.0-alpha.1...mini-rx-store-5.0.0-alpha.2) (2023-02-03)
+
+## [5.0.0-alpha.1](https://github.com/spierala/mini-rx-store/compare/mini-rx-store-5.0.0-alpha.0...mini-rx-store-5.0.0-alpha.1) (2023-02-01)
+
+
+### Features
+
+* **mini-rx-store:** Component Store: dispatch a destroy Action (for logging) on destroy ([d678548](https://github.com/spierala/mini-rx-store/commit/d678548ff9a708fdc56e8ec4a8897ed1a7bbfb88))
+
+## [5.0.0-alpha.0](https://github.com/spierala/mini-rx-store/compare/mini-rx-store-4.2.0...mini-rx-store-5.0.0-alpha.0) (2023-01-23)
+
+
+### ⚠ BREAKING CHANGES
+
+* **mini-rx-store:** configureStore does not return a Store instance, it returns an object
+
+### Features
+
+* **mini-rx-store:** component store ([50bb452](https://github.com/spierala/mini-rx-store/commit/50bb452a8beb1323323721c829cc7dfc42b4b100))
+* **mini-rx-store:** lazy state initialisation (setInitialState),
+* **mini-rx-store:** setState with Observable
+* **mini-rx-store:** more tree-shakable
+
 ## [4.2.0](https://github.com/spierala/mini-rx-store/compare/mini-rx-store-4.1.0...mini-rx-store-4.2.0) (2022-11-24)